@@ -2,21 +2,38 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <!--<TargetFrameworks>net6.0;netcoreapp3.1;net35</TargetFrameworks>-->
-	<TargetFrameworks>net6.0</TargetFrameworks>
-	<RootNamespace>HiBiKiRadioTool.Console</RootNamespace>
+    <TargetFrameworks>net7.0;net6.0;net481;net472;net35</TargetFrameworks>
+	  <RootNamespace>HiBiKiRadioTool.Console</RootNamespace>
     <AssemblyName>hbkr</AssemblyName>
-	<LangVersion>latestMajor</LangVersion>
-	<Nullable>enable</Nullable>
-	<DebugType>embedded</DebugType>
+	  <LangVersion>latestMajor</LangVersion>
+	  <Nullable>enable</Nullable>
+	  <DebugType>embedded</DebugType>
+    <NoWarn Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net7.0'))">SYSLIB1045</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Core\HiBiKiRadio.Core\HiBiKiRadio.Core.csproj" />
+    <Using Include="System" />
+    <Using Include="System.Collections.Generic" />
+    <Using Include="System.IO" />
+    <Using Include="System.Linq" />
+    <Using Include="System.Text" />
+    <Using Include="System.Text.RegularExpressions" />
+    <Using Include="System.Threading" />
+    <Using Include="System.Threading.Tasks" />
   </ItemGroup>
 
-  <Import Condition="'$(TargetFramework)' == 'netstandard2.0'" Project="..\..\Shared\NullableAttributes\NullableAttributes.projitems" Label="Shared" />
+  <ItemGroup Condition="'$(TargetFramework)' == 'net35'">
+    <PackageReference Include="TaskParallelLibrary" Version="*" />
+  </ItemGroup>
 
-  <Import Condition="'$(TargetFramework)' == 'net35'" Project="..\..\Shared\NullableAttributes\NullableAttributes.projitems" Label="Shared" />
+  <ItemGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net45'))">
+    <ProjectReference Include="..\..\Core\HiBiKiRadio.Core\HiBiKiRadio.Core.csproj" SetTargetFramework="TargetFramework=netstandard2.0" />
+  </ItemGroup>
+
+  <ItemGroup Condition="!$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net45'))">
+	  <ProjectReference Include="..\..\Core\HiBiKiRadio.Core\HiBiKiRadio.Core.csproj" />
+  </ItemGroup>
+
+  <Import Project="..\..\Shared\NullableAttributes\NullableAttributes.projitems" Label="Shared" />
 
 </Project>
