@@ -1,16 +1,16 @@
-﻿namespace SamLu.Utility.HiBiKiRadio.Json
+﻿namespace SamLu.Utility.HiBiKiRadio.Json;
+
+#pragma warning disable CS1591
+public class cast
 {
-    public class cast
-    {
-        public int id { get; set; }
-        public string name { get; set; }
-        public string roll_name { get; set; }
-        public string pc_image_url { get; set; }
-        public image_info pc_image_info { get; set; }
-        public string sp_image_url { get; set; }
-        public image_info sp_image_info { get; set; }
-        public string publish_start_at { get; set; }
-        public string publish_end_at { get; set; }
-        public string updated_at { get; set; }
-    }
+    public int id { get; set; }
+    public string? name { get; set; }
+    public string? roll_name { get; set; }
+    public string? pc_image_url { get; set; }
+    public image_info? pc_image_info { get; set; }
+    public string? sp_image_url { get; set; }
+    public image_info? sp_image_info { get; set; }
+    public string? publish_start_at { get; set; }
+    public string? publish_end_at { get; set; }
+    public string? updated_at { get; set; }
 }
