@@ -1,8 +1,8 @@
-﻿namespace SamLu.Utility.HiBiKiRadio.Json
+﻿namespace SamLu.Utility.HiBiKiRadio.Json;
+
+#pragma warning disable CS1591
+public class image_info
 {
-    public class image_info
-    {
-        public int width { get; set; }
-        public int height { get; set; }
-    }
+    public int width { get; set; }
+    public int height { get; set; }
 }
