@@ -1,12 +1,12 @@
-﻿namespace SamLu.Utility.HiBiKiRadio.Json
+﻿namespace SamLu.Utility.HiBiKiRadio.Json;
+
+#pragma warning disable CS1591
+public class information_part
 {
-    public class information_part
-    {
-        public int sort_order { get; set; }
-        public string description { get; set; }
-        public string pc_image_url { get; set; }
-        public image_info pc_image_info { get; set; }
-        public string sp_image_url { get; set; }
-        public image_info sp_image_info { get; set; }
-    }
+    public int sort_order { get; set; }
+    public string? description { get; set; }
+    public string? pc_image_url { get; set; }
+    public image_info? pc_image_info { get; set; }
+    public string? sp_image_url { get; set; }
+    public image_info? sp_image_info { get; set; }
 }
