@@ -1,8 +1,8 @@
-﻿namespace SamLu.Utility.HiBiKiRadio.Json
+﻿namespace SamLu.Utility.HiBiKiRadio.Json;
+
+#pragma warning disable CS1591
+public class playlist
 {
-    public class playlist
-    {
-        public string playlist_url { get; set; }
-        public string token { get; set; }
-    }
+    public string? playlist_url { get; set; }
+    public string? token { get; set; }
 }
