@@ -1,13 +1,13 @@
-﻿namespace SamLu.Utility.HiBiKiRadio.Json
+﻿namespace SamLu.Utility.HiBiKiRadio.Json;
+
+#pragma warning disable CS1591
+public class program_link
 {
-    public class program_link
-    {
-        public int id{ get; set; }
-        public string name{ get; set; }
-        public string pc_image_url{ get; set; }
-        public image_info pc_image_info{ get; set; }
-        public string sp_image_url{ get; set; }
-        public image_info sp_image_info{ get; set; }
-        public string link_url{ get; set; }
-    }
+    public int id{ get; set; }
+    public string? name{ get; set; }
+    public string? pc_image_url{ get; set; }
+    public image_info? pc_image_info{ get; set; }
+    public string? sp_image_url{ get; set; }
+    public image_info? sp_image_info{ get; set; }
+    public string? link_url{ get; set; }
 }
