@@ -1,13 +1,13 @@
-﻿namespace SamLu.Utility.HiBiKiRadio.Json
+﻿namespace SamLu.Utility.HiBiKiRadio.Json;
+
+#pragma warning disable CS1591
+public class segment
 {
-    public class segment
-    {
-        public int id{ get; set; }
-        public string name{ get; set; }
-        public segment_part[] segment_parts{ get; set; }
-        public string html_description{ get; set; }
-        public string publish_start_at{ get; set; }
-        public string publish_end_at{ get; set; }
-        public string updated_at{ get; set; }
-    }
+    public int id{ get; set; }
+    public string? name{ get; set; }
+    public segment_part[]? segment_parts{ get; set; }
+    public string? html_description{ get; set; }
+    public string? publish_start_at{ get; set; }
+    public string? publish_end_at{ get; set; }
+    public string? updated_at{ get; set; }
 }
