@@ -1,20 +1,25 @@
-namespace System.Diagnostics.CodeAnalysis
+#if !NETCOREAPP3_0_OR_GREATER && !NETSTANDARD2_1_OR_GREATER
+
+namespace System.Diagnostics.CodeAnalysis;
+
+#if DEBUG
+/// <summary>
+/// Specifies that a method that will never return under any circumstance.
+/// </summary>
+#endif
+[AttributeUsage(
+    AttributeTargets.Method,
+    Inherited = false)]
+[DebuggerNonUserCode]
+internal sealed class DoesNotReturnAttribute : Attribute
 {
 #if DEBUG
     /// <summary>
-    ///     Specifies that a method that will never return under any circumstance.
+    /// Initializes a new instance of the <see cref="DoesNotReturnAttribute"/> class.
     /// </summary>
+    ///
 #endif
-    [AttributeUsage(AttributeTargets.Method, Inherited = false)]
-    [DebuggerNonUserCode]
-    internal sealed class DoesNotReturnAttribute : Attribute
-    {
-#if DEBUG
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="DoesNotReturnAttribute"/> class.
-        /// </summary>
-        ///
-#endif
-        public DoesNotReturnAttribute() { }
-    }
+    public DoesNotReturnAttribute() { }
 }
+
+#endif
