@@ -2,29 +2,28 @@
 using Microsoft.Build.Utilities;
 using SamLu.Utility.HiBiKiRadio.Tasks;
 
-namespace SamLu.Utility.HiBiKiRadio.Build.Tasks
+namespace SamLu.Utility.HiBiKiRadio.Build.Tasks;
+
+public class DownloadProgramAudio : Task
 {
-    public class DownloadProgramAudio : Task
+    [Required]
+    public ITaskItem ProgramId { get; set; }
+
+    [Output]
+    public ITaskItem ProgramName { get; protected set; }
+
+    [Output]
+    public ITaskItem EpisodeName { get; protected set; }
+
+    public override bool Execute()
     {
-        [Required]
-        public ITaskItem ProgramId { get; set; }
+        ProgramDetailTask programDetailTask = new();
+        var program = programDetailTask.Fetch(this.ProgramId.ItemSpec);
 
-        [Output]
-        public ITaskItem ProgramName { get; protected set; }
-
-        [Output]
-        public ITaskItem EpisodeName { get; protected set; }
-
-        public override bool Execute()
-        {
-            ProgramDetailTask programDetailTask = new ProgramDetailTask();
-            var program = programDetailTask.Fetch(this.ProgramId.ItemSpec);
-
-            this.ProgramName = new TaskItem(program.Name);
-            this.EpisodeName = new TaskItem(program.Episode.Name);
+        this.ProgramName = new TaskItem(program.Name);
+        this.EpisodeName = new TaskItem(program.Episode.Name);
 
 
-            throw new System.NotImplementedException();
-        }
+        throw new System.NotImplementedException();
     }
 }
