@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿namespace SamLu.Utility.HiBiKiRadio.M3U8;
 
-namespace SamLu.Utility.HiBiKiRadio.M3U8
+public enum InsectionType
 {
-    public enum InsectionType
-    {
-        Comment,
-        Tag,
-        Uri
-    }
+    Comment,
+    Tag,
+    Uri
 }
