@@ -1,20 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿namespace SamLu.Utility.HiBiKiRadio.M3U8;
 
-namespace SamLu.Utility.HiBiKiRadio.M3U8
+public enum M3U8Tag
 {
-    public enum M3U8Tag
-    {
-        EXTM3U,
-        EXT_X_VERSION,
-        EXT_X_MEDIA_SEQUENCE,
-        EXT_X_ALLOW_CACHE,
-        EXT_X_TARGETDURATION,
-        EXT_X_KEY,
-        EXTINF,
-        EXT_X_ENDLIST,
-        EXT_X_DISCONTINUITY
-    }
+    EXTM3U,
+    EXT_X_VERSION,
+    EXT_X_MEDIA_SEQUENCE,
+    EXT_X_ALLOW_CACHE,
+    EXT_X_TARGETDURATION,
+    EXT_X_KEY,
+    EXTINF,
+    EXT_X_ENDLIST,
+    EXT_X_DISCONTINUITY
 }
